Update so U gate uses f64 precision

+++ src/circuit.rs
@@ -74,14 +74,14 @@ impl Circuit {
         self
     }
 
-    pub fn u(mut self, theta: f32, phi: f32, lambda: f32, target: usize) -> Self {
+    pub fn u(mut self, theta: f64, phi: f64, lambda: f64, target: usize) -> Self {
         self.instructions.push(Instruction::Gate(
             Gate::new(GateType::U(theta, phi, lambda), &[], &[target]).unwrap(),
         ));
         self
     }
 
-    pub fn cu(mut self, theta: f32, phi: f32, lambda: f32, control: usize, target: usize) -> Self {
+    pub fn cu(mut self, theta: f64, phi: f64, lambda: f64, control: usize, target: usize) -> Self {
         self.instructions.push(Instruction::Gate(
             Gate::new(GateType::U(theta, phi, lambda), &[control], &[target]).unwrap(),
         ));
